+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **[BREAKING CHANGE]** `<Text>`: changed export for `TextDisplayType` `TextTagType`
 - **[BREAKING CHANGE]** `<Tabs>`: changed export for `TabsStatus`
 - **[BREAKING CHANGE]** `Button`: changed export for `Button.STATUS` to new `ButtonStatus`
+- **[BREAKING CHANGE]** `DatePicker`: changed export of `DatePicker.constants` to `DatePickerOrientation`
 
 # v8.8.0 (24/07/2019)